requirements.txt
schema
pyyaml
yamllint
pandas
graphviz
xmltodict
wordcloud
tomli; python_version < '3.11'
argparse
